requirements.txt
torch
torchvision
opencv-python
scipy
fastai==1.0.52
wandb
tensorboardX==1.6
ffmpeg
ffmpeg-python==0.1.17
youtube-dl>=2019.4.17
jupyterlab
opencv-python>=3.3.0.10
pillow==6.2.2
dlib
scikit-image
easydict
PyYAML
dominate>=2.3.1
dill
